refactored the pipelines.  moved them to their own package
