+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pipeline/GripPipelinePurplePixelRGB.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.pipeline;
 
 import java.io.File;
 import java.io.FileWriter;
