+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pipeline/GripPipelineTwoRedRGB.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.pipeline;
 
 import java.io.File;
 import java.io.FileWriter;
@@ -14,16 +14,15 @@ import org.opencv.core.Core.*;
 import org.opencv.imgcodecs.Imgcodecs;
 import org.opencv.imgproc.*;
 import org.opencv.objdetect.*;
-import org.openftc.easyopencv.OpenCvPipeline;
 
 /**
-* GripPipelineTwoBlueRGB class.
+* GripPipelineTwoRedRGB class.
 *
 * <p>An OpenCV pipeline generated by GRIP.
 *
 * @author GRIP
 */
-class GripPipelineTwoBlueRGB  extends OpenCvPipeline {
+public class GripPipelineTwoRedRGB {
 
 	//Outputs
 	private Mat cvResizeOutput = new Mat();
@@ -40,7 +39,6 @@ class GripPipelineTwoBlueRGB  extends OpenCvPipeline {
 	 * This is the primary method that runs the entire pipeline and updates the outputs.
 	 */
 	public Mat processFrame(Mat source0) {
-		System.out.println("processing requested");
 		// Step CV_resize0:
 		Mat cvResizeSrc = source0;
 		Size cvResizeDsize = new Size(0, 0);
@@ -51,9 +49,9 @@ class GripPipelineTwoBlueRGB  extends OpenCvPipeline {
 
 		// Step RGB_Threshold0:
 		Mat rgbThresholdInput = cvResizeOutput;
-		double[] rgbThresholdRed = {0.0, 182.47439719711147};
-		double[] rgbThresholdGreen = {0.0, 255.0};
-		double[] rgbThresholdBlue = {177.33813368802447, 255.0};
+		double[] rgbThresholdRed = {195.68344157376734, 255.0};
+		double[] rgbThresholdGreen = {0.0, 153.46416935578955};
+		double[] rgbThresholdBlue = {44.3345321918563, 255.0};
 		rgbThreshold(rgbThresholdInput, rgbThresholdRed, rgbThresholdGreen, rgbThresholdBlue, rgbThresholdOutput);
 
 		// Step CV_erode0:
